separate the implementation from the plugin registration
